add xpore-postprocessing
